feat: Added bubble sort and divided out the functions
